+++ apps/shared/avifutil.h
@@ -1,0 +1,6 @@
+// Copyright 2019 Joe Drago. All rights reserved.
+// SPDX-License-Identifier: BSD-2-Clause
+
+#include "avif/avif.h"
+
+void avifImageDump(avifImage * avif);