+++ apps/shared/y4m.h
@@ -1,0 +1,7 @@
+// Copyright 2019 Joe Drago. All rights reserved.
+// SPDX-License-Identifier: BSD-2-Clause
+
+#include "avif/avif.h"
+
+avifBool y4mRead(avifImage * avif, const char * inputFilename);
+avifBool y4mWrite(avifImage * avif, const char * outputFilename);